Split SvgGlyph
